pyproject.toml
[project]
name = "bgp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "google-genai>=1.59.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]
[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "F", "N", "W", "E"]
ignore = ["E501"]